# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 'Application API: bank-rest'
  description: API for bank-rest.
  license:
    name: Apache 2.0
    url: http://springdoc.org
  version: 0.0.1-alpha
servers:
  - url: /api
    description: Current service
paths:
  /api/users:
    put:
      tags:
        - user-controller
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
  /api/cards/status/request-block:
    put:
      tags:
        - card-controller
      operationId: requestBlock
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardDto'
  /api/cards/admin/set-status:
    put:
      tags:
        - card-controller
      operationId: adminUpdateCardStatus
      parameters:
        - name: cardId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardDto'
  /api/cards/admin/block:
    put:
      tags:
        - card-controller
      operationId: blockCard
      parameters:
        - name: cardId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardDto'
  /auth/sign-up:
    post:
      tags:
        - Authentication
      summary: Регистрация нового пользователя
      operationId: signUp
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JwtAuthenticationResponse'
  /auth/sign-in:
    post:
      tags:
        - Authentication
      summary: Аутенификация пользователя
      operationId: signIn
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JwtAuthenticationResponse'
  /api/users/register:
    post:
      tags:
        - user-controller
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
  /api/transfer:
    post:
      tags:
        - transfer-controller
      operationId: transfer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequestDto'
        required: true
      responses:
        '200':
          description: OK
  /api/cards:
    post:
      tags:
        - card-controller
      operationId: createCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardDto'
  /api/cards/transfer:
    post:
      tags:
        - card-controller
      operationId: transfer_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequestDto'
        required: true
      responses:
        '200':
          description: OK
  /api/users/{id}:
    get:
      tags:
        - user-controller
      operationId: getById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
    delete:
      tags:
        - user-controller
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
  /api/users/me:
    get:
      tags:
        - user-controller
      operationId: getCurrentUser
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDto'
  /api/cards/{id}:
    get:
      tags:
        - card-controller
      operationId: getById_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CardDto'
    delete:
      tags:
        - card-controller
      operationId: deleteCard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
  /api/cards/me/paged:
    get:
      tags:
        - card-controller
      operationId: getMyCardsPaged
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 5
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageCardDto'
  /api/cards/admin/all:
    get:
      tags:
        - card-controller
      operationId: getAllCards
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardDto'
components:
  schemas:
    UserDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
    CardDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        maskedNumber:
          type: string
        status:
          type: string
        expireAt:
          type: string
        balance:
          type: number
          format: double
        ownerUsername:
          type: string
    SignUpRequest:
      type: object
      description: Sign Up Request DTO
      properties:
        username:
          type: string
          description: Username
          example: User123
          maxLength: 50
          minLength: 5
        email:
          type: string
          format: email
          description: Email
          example: User123@gmail.com
          maxLength: 255
          minLength: 5
        password:
          type: string
          description: Password
          example: my_1secret1_password
          maxLength: 255
          minLength: 0
      required:
        - email
        - username
    JwtAuthenticationResponse:
      type: object
      description: Respond with JWT token
      properties:
        token:
          type: string
          description: JWT Access token
          example: eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJhZG1pbiIsImV4cCI6MTYyMjUwNj...
    SignInRequest:
      type: object
      description: Request for user login
      properties:
        username:
          type: string
          description: Username
          example: User123
          maxLength: 50
          minLength: 5
        password:
          type: string
          description: Password
          example: my_1secret1_password
          maxLength: 255
          minLength: 8
      required:
        - password
        - username
    CreateUserDto:
      type: object
      properties:
        username:
          type: string
          maxLength: 30
          minLength: 3
        email:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          maxLength: 2147483647
          minLength: 6
        role:
          type: string
          enum:
            - ROLE_ADMIN
            - ROLE_USER
      required:
        - email
        - password
        - role
        - username
    TransferRequestDto:
      type: object
      properties:
        fromCardId:
          type: integer
          format: int64
        toCardId:
          type: integer
          format: int64
        amount:
          type: number
          format: double
          minimum: 1
      required:
        - amount
        - fromCardId
        - toCardId
    CreateCardDto:
      type: object
      properties:
        expireAt:
          type: string
          pattern: \d{4}-\d{2}-\d{2}
        balance:
          type: number
          format: double
          minimum: 0
        username:
          type: string
          maxLength: 30
          minLength: 3
      required:
        - balance
        - expireAt
        - username
    PageCardDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        numberOfElements:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/CardDto'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        paged:
          type: boolean
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
    SortObject:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean